+++ include/inc_lib/revision/revision.php
@@ -9,8 +9,8 @@
  *
  **/
 
-define('PHPWCMS_VERSION',		'1.6.519');
-define('PHPWCMS_RELEASE_DATE',	'2012/11/22');
-define('PHPWCMS_REVISION',		'519');
+define('PHPWCMS_VERSION',		'1.6.520');
+define('PHPWCMS_RELEASE_DATE',	'2012/11/23');
+define('PHPWCMS_REVISION',		'520');
 
 ?>