+++ CHANGES.md
@@ -2,6 +2,17 @@ phpwcms Changelog
 =======
 
 
+Version 1.6.520 – November 23, 2012
+-----------------------------------
+
+- Fix backend file manager download
+- Some more fixes for file management when logged in as admin user
+- Updated Polish translations (UTF-8), Thanks Bogdan!
+- Updated versions of editors under profile account data
+- Removed mediabox from lib
+- Updated some more file headers with updated copyright info
+
+
 Version 1.6.519 – November 22, 2012
 -----------------------------------